Remove types and move a few files
