+++ src/abi/index.ts
@@ -9,8 +9,6 @@
 
 export { AbiCoder } from './abi-coder.js';
 
-export { decodeBytes32String, encodeBytes32String } from './bytes32.js';
-
 export {
     ConstructorFragment,
     ErrorFragment,
