+++ src/address/index.ts
@@ -35,4 +35,4 @@ export { getAddress, computeAddress, recoverAddress, formatMixedCaseChecksumAddr
 
 export { getCreateAddress, getCreate2Address } from './contract-address.js';
 
-export { isAddressable, isAddress, resolveAddress, validateAddress } from './checks.js';
+export { isAddressable, isAddress, resolveAddress, validateAddress, isQuaiAddress, isQiAddress } from './checks.js';
