+++ src/encoding/index.ts
@@ -1,3 +1,4 @@
+export { decodeBytes32, encodeBytes32 } from './bytes32.js';
 export { decodeBase58, encodeBase58 } from './base58.js';
 export { decodeBase64, encodeBase64 } from './base64.js';
 export { encodeProtoTransaction, encodeProtoWorkObject } from './proto-encode.js';
