+++ src/utils/shards.ts
@@ -1,4 +1,5 @@
 import { toZone, Zone } from '../constants/zones.js';
+import { isQiAddress } from '../quais.js';
 /**
  * Retrieves the shard information for a given address based on its byte prefix. The function parses the address to
  * extract its byte prefix, then filters the ShardData to find a matching shard entry. If no matching shard is found, it
@@ -40,42 +41,22 @@ export function getAddressDetails(address: string): { zone: Zone; isUTXO: boolea
  * Otherwise, it returns 0.
  *
  * @category Utils
- * @param {string | null} from - The sender address, expected to start with "0x" followed by its hexadecimal
- *   representation. If null, the function returns 0.
- * @param {string | null} to - The recipient address, expected to start with "0x" followed by its hexadecimal
- *   representation. If null, the function returns 0.
+ * @param {string | null} from - The sender address. If null, the function returns 0.
+ * @param {string | null} to - The recipient address. If null, the function returns 0.
  *
  * @returns {number} The transaction type based on the addresses.
  */
 export function getTxType(from: string | null, to: string | null): number {
     if (from === null || to === null) return 0;
-    const fromUTXO = isQiAddress(from);
-    const toUTXO = isQiAddress(to);
+    const senderAddressIsQi = isQiAddress(from);
+    const recipientAddressIsQi = isQiAddress(to);
 
     switch (true) {
-        case fromUTXO && toUTXO:
+        case senderAddressIsQi && recipientAddressIsQi:
             return 2;
-        case fromUTXO && !toUTXO:
+        case senderAddressIsQi && !recipientAddressIsQi:
             return 2;
         default:
             return 0;
     }
-}
-
-/**
- * Checks whether a given blockchain address is a UTXO address based on the 9th bit of the address. This function
- * extracts the second byte of the address and checks its first bit to determine the UTXO status.
- *
- * @category Utils
- * @param {string} address - The blockchain address to be analyzed, expected to start with "0x" followed by its
- *   hexadecimal representation.
- *
- * @returns {boolean} True if the address is a UTXO address, false otherwise.
- */
-export function isQiAddress(address: string): boolean {
-    const secondByte = address.substring(4, 6);
-    const binaryString = parseInt(secondByte, 16).toString(2).padStart(8, '0');
-    const isUTXO = binaryString[0] === '1';
-
-    return isUTXO;
 }
